requirements.txt
# --- requirements.txt ---
# Install these using: pip install -r requirements.txt

# Core FastAPI and web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0

# Database and storage
pymongo==4.6.0
redis==5.0.1
fakeredis==2.20.1

# AI and LangChain
langgraph==0.2.11
langchain==0.2.14
langchain-groq==0.1.9

# Authentication and security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1

# Data validation and settings
pydantic==2.5.0
pydantic-settings==2.1.0

# HTTP and API
requests==2.31.0

# Image processing
Pillow==10.1.0

# Web and templates
jinja2==3.1.2
python-multipart==0.0.6
websockets==12.0

# Background tasks
arq==0.25.0
 